Put fromRqid on game.system.api Fixes #227
- added src\system\api directory
- added rqidApi.ts
- made fromRqid a property of game.system.api
- fixed references to the old getItemByRqid to use fromRqid
- fixed "Get Item Like this" macro in RQG System
- (unrelated) If occupational skill gives a 0% bonus it reads as "+0%" in display mode (instead of "-0%"
