+++ src/items/sheet-parts/itemRqgSystem.hbs
@@ -99,7 +99,7 @@
       type="text"
       id="get-item-like-this-macro-{{data._id}}"
       name="data.get-item-like-this-macro"
-      value='await CONFIG.Item.documentClass.getItemByRqid("{{itemData.rqid}}","{{itemData.rqidLang}}")'
+      value='await game.system.api.fromRqid("{{itemData.rqid}}","{{itemData.rqidLang}}")'
       readonly
     />
     <a title="{{localize "RQG.RQGSystem.CopyInputValue"}}"
