+++ src/rqg.ts
@@ -13,6 +13,7 @@ import { RqgToken } from "./combat/rqgToken";
 import { getGame, RqgError } from "./system/util";
 import { consolidateCompendiumItems } from "./system/migrations/ConsolidateItems";
 import { RqgPause } from "./foundryUi/pause";
+import { fromRqid } from "./system/api/rqidApi";
 
 Hooks.once("init", async () => {
   console.log(
@@ -66,6 +67,7 @@ Hooks.once("init", async () => {
     // installModules: installModules,
     migrate: applyDefaultWorldMigrations,
     consolidate: consolidateCompendiumItems,
+    fromRqid: fromRqid,
   };
 });
